edit interfaces and tests

+++ database/interfaces/redis_json.py
@@ -1,11 +1,12 @@
 import warnings
 from abc import ABC
-from typing import Any, Union, Optional
+from typing import Any, Optional
 
-from aredis_om import NotFoundError, Migrator, get_redis_connection, HashModel
+from aredis_om import NotFoundError, get_redis_connection, HashModel, Migrator
+from pydantic import BaseModel
 
 from config import settings
-from .base_interface import BaseDBInterface
+from .base_interface import BaseDBInterface, SchemasValidator
 
 
 class BaseRedisModel(HashModel, ABC):
@@ -14,11 +15,13 @@ class BaseRedisModel(HashModel, ABC):
         database = get_redis_connection(url=settings.REDIS_URL)
 
 
-class BaseRedisInterface(BaseDBInterface):
+class BaseRedisInterface(BaseDBInterface, SchemasValidator):
     _base_schemas = None
+    _filter_schemas = None
 
     def __init__(self,
-                 base_schemas: Union[HashModel, Any]
+                 base_schemas: HashModel.model_json_schema,
+                 filter_schemas: BaseModel.model_json_schema
                  ):
         warnings.filterwarnings(
             "ignore",
@@ -31,12 +34,19 @@ class BaseRedisInterface(BaseDBInterface):
         )
 
         self._base_schemas = base_schemas
+        self._filter_schemas = filter_schemas
+
+    async def migrate(self):
+        await Migrator().run()
+        return self
 
     async def __connect_filter_with_kwargs(self, where_filter: Any = None,
                                            error: bool = True, **kwargs):
         if error:
             if where_filter is None and not kwargs:
                 raise ValueError('At least one of `where_filter` or `kwargs` must be set')
+
+        kwargs = await self.valid_schema(self._filter_schemas, **kwargs)
 
         for key, value in kwargs.items():
             model_item = getattr(self._base_schemas, key)
@@ -47,16 +57,20 @@ class BaseRedisInterface(BaseDBInterface):
 
         return where_filter
 
-    async def migrate(self):
-        await Migrator().run()
-        return self
+    async def __create(self, create_object: _base_schemas):
+        try:
+            await self._delete(self._base_schemas.id == create_object.id)
+        except:
+            pass
+        await create_object.save()
+        return True
 
     async def _create(self, create_object: _base_schemas | list[_base_schemas]) -> bool:
         if isinstance(create_object, list):
             for item in create_object:
-                await item.save()
+                await self.__create(item)
         else:
-            await create_object.save()
+            await self.__create(create_object)
 
         return True
 
@@ -64,19 +78,23 @@ class BaseRedisInterface(BaseDBInterface):
         try:
             where_filter = await self.__connect_filter_with_kwargs(where_filter, **kwargs)
             return await self._base_schemas.find(where_filter).first()
-
         except NotFoundError:
             return None
 
     async def _get_all(self,
                        where_filter: Any = None,
-                       limit: int = 10000,
+                       limit: int = 10,
                        offset: int = 0,
-                       **kwargs) -> Any:
+                       **kwargs) -> list[_base_schemas]:
         try:
             where_filter = await self.__connect_filter_with_kwargs(where_filter,
                                                                    error=False,
                                                                    **kwargs)
+
+            if where_filter:
+                where_filter = where_filter & (self._base_schemas.id >= offset)
+            else:
+                where_filter = self._base_schemas.id >= offset
 
             if where_filter:
                 query = self._base_schemas.find(where_filter)
@@ -84,13 +102,29 @@ class BaseRedisInterface(BaseDBInterface):
                 query = self._base_schemas.find()
 
             query.limit = limit
-            query.offset = offset
+            query.offset = 0
 
             return await query.all()
         except NotFoundError:
-            return None
+            return []
+
+    async def __update(self, update_object: _base_schemas):
+        del_res = await self._delete(self._base_schemas.id == update_object.id)
+        if del_res == False:
+            return False
+        await self._create(update_object)
+        return True
+
+    async def _update(self, update_object: _base_schemas | list[_base_schemas]) -> bool:
+        if isinstance(update_object, list):
+            for item in update_object:
+                await self.__update(item)
+        else:
+            await self.__update(update_object)
 
     async def _delete(self, where_filter: Any) -> bool:
         models = await self._get_all(where_filter)
+        if not models:
+            return False
         await self._base_schemas.delete_many(models)
         return True
